backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM & AI
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16
openai==1.10.0
langchain-text-splitters==0.1.1

# Vector Database
chromadb==0.4.22
pinecone-client==3.0.2

# Embeddings
sentence-transformers==2.3.1

# Caching & Storage
redis==5.0.1
aioredis==2.0.1

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
pyyaml==6.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Monitoring
prometheus-client==0.19.0

# PDF Processing (for policy documents)
pypdf==3.17.4
